feat: contact form component
all code has been added into the component. the form data gets send to the +page.server.js
